Add Mineswapper2 tab with adjustable board sizes
